feat: resolves requested changes
