+++ data_structures/hashing/hash_map.ts
@@ -1,17 +1,4 @@
-/**
- * This interface is a representation of the Map data structure.
- */
-export interface Map<K, V> {
-    getSize(): number;
-    set(key: K, value: V): void;
-    get(key: K): V | null;
-    delete(key: K): void;
-    has(key: K): boolean;
-    clear(): void;
-    keys(): K[];
-    values(): V[];
-    entries(): HashMapEntry<K, V>[];
-}
+import { Map } from "../map";
 
 /**
  * Represents a hash map.
@@ -239,7 +226,7 @@ export class HashMap<K, V> implements Map<K, V> {
  * @param key The key.
  * @param value The value.
  */
-class HashMapEntry<K, V> {
+export class HashMapEntry<K, V> {
     key: K;
     value: V;
